Fix some issue with labels read as numeric
In case the label of a component, connection or bus was a string
representation of a numerical value (e.g. "1") in script used to
export the network information, the network reader tried to read
the data and interpret them as numerical value instead of a string.

Furthermore some methods have been renamed for clarification.