Add a second draft component
The `MovingBoundaryHeatExchanger` is the more generic version of
the first implementation. It allows for any number of sections and
eventually will also be implemented for evaporation.

The main difference to the first draft of the
`MovingBoundaryCondenser` is, that it is nondimensional to U or A.
The sum of all UA values in the sections has to be equal to the UA
value originally calculated in the design phase. An adoption can
be made to this later, modifying the total UA value by lookup
tables similar to the ´kA_char´ approaches in the already existing
heat exchanger components.